backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.2
httpx==0.27.0
requests==2.32.5
boto3==1.34.136
python-multipart==0.0.9
python-json-logger==2.0.7
prometheus_client==0.20.0
python-dotenv==1.0.1
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
opentelemetry-propagator-aws-xray==1.0.2
opentelemetry-sdk-extension-aws==2.1.0
opentelemetry-distro==0.59b0